feat: split trpc backend into files, create config folder inside client
