+++ server/src/index.ts
@@ -1,12 +1,13 @@
-import * as trpcExpress from '@trpc/server/adapters/express';
 import cors from 'cors';
 import express from 'express';
 
-import { appRouter } from './trpc';
+import { applyTrpcToExpressApp } from './lib/express';
+import { trpcBackendRouter } from './router';
 
 const expressApp = express();
 expressApp.use(cors());
-expressApp.use('/api', trpcExpress.createExpressMiddleware({ router: appRouter }));
+
+applyTrpcToExpressApp(expressApp, trpcBackendRouter);
 
 expressApp.listen('3000', () => {
   console.info('Listening on port 3000');
