+++ server/src/lib/spells.ts
@@ -1,0 +1,17 @@
+export const spells = [
+  {
+    spellName: 'revelio',
+    book: 'Book 1',
+    description: 'Description 1',
+  },
+  {
+    spellName: 'flipendo',
+    book: 'Book 2',
+    description: 'Description 2',
+  },
+  {
+    spellName: 'alohomora',
+    book: 'Book 3',
+    description: 'Description 3',
+  },
+];
