+++ server/src/router/getSpells/index.ts
@@ -1,0 +1,6 @@
+import { spells } from '../../lib/spells';
+import { publicProcedure } from '../../lib/trpc';
+
+export const getSpellsTrpcRoute = publicProcedure.query(() => {
+  return { spells };
+});
