+++ server/src/router/index.ts
@@ -1,0 +1,11 @@
+import { router } from '../lib/trpc';
+
+import { getSpellTrpcRoute } from './getSpell';
+import { getSpellsTrpcRoute } from './getSpells';
+
+export const trpcBackendRouter = router({
+  getSpells: getSpellsTrpcRoute,
+  getSpell: getSpellTrpcRoute,
+});
+
+export type TrpcBackendRouter = typeof trpcBackendRouter;
